Modulaizando filtros, parte 1(Falta modularizar la funciones de los eventos 'change' para los imputs de ambos filtros)
